+++ mek_data_class/lib/src/data_collection_equality.dart
@@ -1,14 +1,16 @@
 import 'package:collection/collection.dart';
 
-class DataCollectionEquality implements Equality<Object?> {
-  const DataCollectionEquality();
+class $CollectionEquality implements Equality<Object?> {
+  final Equality<dynamic> base;
+
+  const $CollectionEquality([this.base = const DefaultEquality<Never>()]);
 
   @override
-  bool equals(e1, e2) {
+  bool equals(Object? e1, Object? e2) {
     if (e1 is List) return e2 is List && ListEquality(this).equals(e1, e2);
     if (e1 is Map) return e2 is Map && MapEquality(keys: this, values: this).equals(e1, e2);
     if (e1 is Set) return e2 is Set && SetEquality(this).equals(e1, e2);
-    return e1 == e2;
+    return base.equals(e1, e2);
   }
 
   @override
@@ -16,7 +18,7 @@ class DataCollectionEquality implements Equality<Object?> {
     if (o is List) return ListEquality(this).hash(o);
     if (o is Map) return MapEquality(keys: this, values: this).hash(o);
     if (o is Set) return SetEquality(this).hash(o);
-    return o.hashCode;
+    return base.hash(o);
   }
 
   @override